Get stuff out of the companion object
