+++ src/main/kotlin/org/kamiblue/botkt/config/GlobalConfigs.kt
@@ -1,0 +1,80 @@
+package org.kamiblue.botkt.config
+
+import org.kamiblue.botkt.Main
+import org.kamiblue.botkt.event.BotEventBus
+import org.kamiblue.botkt.event.events.ShutdownEvent
+import org.kamiblue.event.listener.listener
+
+object GlobalConfigs {
+    private val globalConfigs = LinkedHashSet<GlobalConfig>()
+
+    init {
+        listener<ShutdownEvent> {
+            saveAll()
+        }
+
+        BotEventBus.subscribe(this)
+    }
+
+    fun register(config: GlobalConfig) {
+        config.load()
+        globalConfigs.add(config)
+    }
+
+    fun unregister(config: GlobalConfig) {
+        globalConfigs.remove(config)
+    }
+
+    fun loadAll(): Boolean {
+        return globalConfigs.runCatchingAll(GlobalConfig::load) { it, e ->
+            Main.logger.warn("Failed to load global config ${it.name}", e)
+        }
+    }
+
+    fun saveAll(): Boolean {
+        return globalConfigs.runCatchingAll(GlobalConfig::save) { it, e ->
+            Main.logger.warn("Failed to save global config ${it.name}", e)
+        }
+    }
+
+    fun load(config: GlobalConfig): Boolean {
+        return try {
+            checkRegistered(config)
+            config.load()
+            true
+        } catch (e: Exception) {
+            Main.logger.warn("Failed to load global config ${config.name}", e)
+            false
+        }
+    }
+
+    fun save(config: GlobalConfig): Boolean {
+        return try {
+            checkRegistered(config)
+            config.save()
+            true
+        } catch (e: Exception) {
+            Main.logger.warn("Failed to save global config ${config.name}", e)
+            false
+        }
+    }
+
+    private fun checkRegistered(config: GlobalConfig) {
+        if (!globalConfigs.contains(config)) {
+            throw IllegalArgumentException("Global config type ${config.name} is not registered")
+        }
+    }
+
+    private fun <T> Iterable<T>.runCatchingAll(block: (T) -> Unit, catch: (T, Exception) -> Unit): Boolean {
+        var success = true
+        this.forEach {
+            try {
+                block(it)
+            } catch (e: Exception) {
+                success = false
+                catch(it, e)
+            }
+        }
+        return success
+    }
+}
